Move CExtraTime traits
